requirements.txt
distro
google-api-python-client
pandas
peewee
pendulum
praw
prettytable
psutil
publicsuffixlist
pymysql
py-cpuinfo
requests
schedule
slackclient==1.3.1
